make an insert query on workerthread to retrieve id from new property freshly created with love to update PhotoEntity.kt

+++ app/src/main/java/com/kardabel/realestatemanager/ui/properties/PropertyViewState.kt
@@ -1,7 +1,7 @@
 package com.kardabel.realestatemanager.ui.properties
 
 data class PropertyViewState(
-    val id: Int,
+    val id: Long,
     val type: String?,
     val address: String?,
     val price: String?
